+++ src/SIFTS/XMLParser.jl
@@ -9,12 +9,9 @@ The extension of the downloaded file is `.xml.gz` by default.
 The `filename` can be changed, but the `.xml.gz` at the end is mandatory.
 """
 function downloadsifts(pdbcode::String; filename::String="$(lowercase(pdbcode)).xml.gz")
-    @assert endswith(filename,".xml.gz") "filename must end with .xml.gz"
+    @assert endswith(filename, ".xml.gz") "filename must end with .xml.gz"
     if check_pdbcode(pdbcode)
-        # Using download instead of download_file because Requests.jl doesn't support ftp:
-        # https://github.com/JuliaWeb/Requests.jl/issues/89
-        download(string(
-            "ftp://ftp.ebi.ac.uk/pub/databases/msd/sifts/split_xml/",
+        download_file(string("ftp://ftp.ebi.ac.uk/pub/databases/msd/sifts/split_xml/",
             lowercase(pdbcode[2:3]), "/", lowercase(pdbcode), ".xml.gz"), filename)
     else
         throw(ErrorException("$pdbcode is not a correct PDB"))
@@ -42,8 +39,8 @@ WARNING: Sometimes there are more chains than entities!
 ```
 """
 function _get_entities(sifts)
-	siftsroot = root(sifts)
-	get_elements_by_tagname(siftsroot, "entity")
+    	siftsroot = root(sifts)
+    	get_elements_by_tagname(siftsroot, "entity")
 end
 
 """
@@ -65,7 +62,7 @@ Returns an Iterator of the residues on the listResidue
 ```
 """
 function _get_residues(segment)
-    child_elements(select_element(get_elements_by_tagname(segment,"listResidue"),"listResidue"))
+    child_elements(select_element(get_elements_by_tagname(segment, "listResidue"), "listResidue"))
 end
 
 """